Add EULA enforcement and faction support initialization
This update enforces EULA acceptance by disabling the plugin if the 'eula' setting in `config.yml` is not set to true. It also adds initialization support for factions through the `HelixFactions` class and improves logging consistency by replacing the custom logger with the plugin-provided logger. Additional functionalities include saving the default config and loading faction variables in the `DataManager`.